sql/pgwire: Extend generate-decimal program for other types
This commit extends the generate-decimal program to work for other
types, renaming the program to generate-binary along the way. It now
accepts `decimal`, `timestamp`, and `timestamptz`, but other types
are simple to add.
