+++ sql/pgwire/testdata/timestamp_test.json
@@ -1,0 +1,26 @@
+[
+	{
+		"In": "1999-01-08 04:05:06",
+		"Expect": [0, 0, 0, 8, 255, 255, 227, 225, 177, 91, 128, 128]
+	},
+	{
+		"In": "2004-10-19 10:23:54",
+		"Expect": [0, 0, 0, 8, 0, 0, 137, 201, 15, 13, 226, 128]
+	},
+	{
+		"In": "0001-01-01 00:00:00",
+		"Expect": [0, 0, 0, 8, 255, 31, 226, 255, 197, 156, 96, 0]
+	},
+	{
+		"In": "0004-10-19 10:23:54",
+		"Expect": [0, 0, 0, 8, 255, 32, 80, 6, 42, 191, 2, 128]
+	},
+	{
+		"In": "4004-10-19 10:23:54",
+		"Expect": [0, 0, 0, 8, 0, 224, 195, 139, 243, 92, 194, 128]
+	},
+	{
+		"In": "9004-10-19 10:23:54",
+		"Expect": [0, 0, 0, 8, 3, 17, 83, 233, 31, 54, 66, 128]
+	}
+]
